Moved settings prompt logic to its own file
